add explicit float/double/long double conversion operators to MPint
Implicit conversions were ruled out: they cause overload ambiguity
when MPint is mixed with integral literals (e.g. a*3 is ambiguous
between MPint::operator* and the built-in double*int).

det.cpp: add one static_cast<long double> at the point where a Bint
matrix entry is copied into the floating-point reference matrix.

+++ src/det.cpp
@@ -197,7 +197,7 @@ static void sameRandomMatrix(const unsigned int dim, const SQState *statePtr,
     perm[ridx] = ridx;
     for (unsigned int cidx = 0; cidx < dim; cidx++) {
       matrix[ridx][cidx] = statePtr->flatMatrix[ridx * dim + cidx];
-      dmatrix[ridx][cidx] = statePtr->flatMatrix[ridx * dim + cidx];
+      dmatrix[ridx][cidx] = static_cast<long double>(statePtr->flatMatrix[ridx * dim + cidx]);
     }
   }
 }
